# Conflicts:
#	HopInNow/.idea/caches/build_file_checksums.ser
#	HopInNow/.idea/caches/gradle_models.ser
#	HopInNow/app/src/main/java/com/example/hopinnow/activities/RiderMapActivity.java
#	HopInNow/app/src/main/java/com/example/hopinnow/activities/RiderPaymentActivity.java

+++ HopInNow/app/src/main/java/com/example/hopinnow/activities/RiderDriverOfferFragment.java
@@ -53,7 +53,7 @@ public class RiderDriverOfferFragment extends Fragment {
             if (driver.getRating()==-1){
                 rating = "not yet rated";
             } else {
-                rating = Integer.toString(driver.getRating());
+                rating = Double.toString(driver.getRating());
             }
             driverRating.setText(rating);
 
@@ -128,7 +128,7 @@ public class RiderDriverOfferFragment extends Fragment {
         if (driver.getRating()==-1){
             rating = "not yet rated";
         } else {
-            rating = Integer.toString(driver.getRating());
+            rating = Double.toString(driver.getRating());
         }
         driverRating.setText(rating);
